+++ player/js/elements/CompElement.js
@@ -51,11 +51,7 @@ ICompElement.prototype.renderFrame = function(parentMatrix){
 
     this.hidden = false;
     for( i = 0; i < len; i+=1 ){
-        if(this.data.hasMask){
-            this.elements[i].renderFrame();
-        }else{
-            this.elements[i].renderFrame(this.finalTransform);
-        }
+        this.elements[i].renderFrame();
     }
     if(this.firstFrame){
         this.firstFrame = false;